backend/python_backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
motor==3.4.0
pymongo==4.7.2
pydantic==2.7.1
PyJWT==2.8.0
httpx==0.27.0
openai==1.30.1
anthropic==0.28.0
google-generativeai==0.5.4
boto3==1.34.110
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==1.1.2
pandas==2.2.2
Pillow==10.3.0
cryptography==42.0.7
aiofiles==23.2.1
websockets==12.0
deepgram-sdk==3.2.7
elevenlabs==1.2.0
requests==2.31.0
gunicorn==22.0.0
